# Use strings.Builder in unpackString