This improves the performance of `unpackString` and reduces the memory allocations.

```
name             old time/op    new time/op    delta
UnpackString-12     186ns ±29%     122ns ±11%  -34.58%  (p=0.000 n=10+10)

name             old alloc/op   new alloc/op   delta
UnpackString-12     80.0B ± 0%     48.0B ± 0%  -40.00%  (p=0.000 n=10+10)

name             old allocs/op  new allocs/op  delta
UnpackString-12      3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.000 n=10+10)
```